Compete in USACO 2021 December, Bronze
points: 667/1000 pts (passed 2/3 problems perfectly)

Round overview:
"The bronze division had 9974 total participants, of whom 7673 were pre-college students. All competitors who scored 700 or higher on this contest (1774 pre-college students) are automatically promoted to the silver division.

The bronze cutoff was set fairly low to account for the fact that the bronze contest this time around did not include any trivial problems (and some problems even included a handful of fun larger cases for a challenge, although solving these was not necessary to achieve a high enough score for promotion)."
=> harder round = tougher to ICP :-/ :-(

Signed-off-by: Karel Ha <mathemage@gmail.com>